# swagger.json
{"openapi":"3.0.0","info":{"title":"Fireworld API","version":"1.0.0","description":"Dokumentacja API dla Fireworld"},"servers":[{"url":"http://localhost:3000"}],"paths":{"/activity":{"get":{"summary":"Pobierz aktywność użytkownika","tags":["Activity"],"responses":{"200":{"description":"Lista aktywności"}}},"post":{"summary":"Dodaj aktywność","tags":["Activity"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"type":{"type":"string"},"description":{"type":"string"}}}}}},"responses":{"201":{"description":"Dodano aktywność"}}}},"/auth/register":{"post":{"summary":"Rejestracja użytkownika","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"201":{"description":"Utworzono użytkownika"},"400":{"description":"Błąd walidacji"}}}},"/auth/login":{"post":{"summary":"Logowanie użytkownika","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Zalogowano pomyślnie"},"401":{"description":"Nieprawidłowe dane"}}}},"/chat/rooms":{"get":{"summary":"Pobierz dostępne pokoje czatu","tags":["Chat"],"responses":{"200":{"description":"Lista pokojów"}}}},"/chat/rooms/{roomId}/join":{"post":{"summary":"Dołącz do pokoju","tags":["Chat"],"parameters":[{"in":"path","name":"roomId","schema":{"type":"string"},"required":true,"description":"ID pokoju"}],"responses":{"200":{"description":"Dołączono"}}}},"/messages":{"post":{"summary":"Wyślij wiadomość","tags":["Messages"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"recipientId":{"type":"string"},"content":{"type":"string"}}}}}},"responses":{"201":{"description":"Wiadomość wysłana"}}}},"/messages/{userId}":{"get":{"summary":"Pobierz wiadomości z użytkownikiem","tags":["Messages"],"parameters":[{"in":"path","name":"userId","schema":{"type":"string"},"required":true,"description":"ID odbiorcy"}],"responses":{"200":{"description":"Lista wiadomości"}}}},"/posts":{"post":{"summary":"Utwórz nowy post","tags":["Posts"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"content":{"type":"string"},"image":{"type":"string"}}}}}},"responses":{"201":{"description":"Post utworzony"}}},"get":{"summary":"Pobierz wszystkie posty","tags":["Posts"],"responses":{"200":{"description":"Lista postów"}}}},"/users/{id}":{"get":{"summary":"Pobierz dane użytkownika","tags":["Users"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Szczegóły użytkownika"}}}},"/users":{"get":{"summary":"Pobierz listę użytkowników","tags":["Users"],"responses":{"200":{"description":"Lista użytkowników"}}}}},"components":{},"tags":[]}